Fix implicit conversion warnings in VariableWidth.cpp

+++ src/libslic3r/VariableWidth.cpp
@@ -125,7 +125,7 @@ static ExtrusionPaths thick_polyline_to_extrusion_paths_2(const ThickPolyline& t
                 double length = lines[start_index].length();
                 double sum = lines[start_index].length() * 0.5 * (lines[start_index].a_width + lines[start_index].b_width);
                 path.polyline.append(lines[start_index].a);
-                for (int idx = start_index + 1; idx < i; idx++) {
+                for (int idx = start_index + 1; idx < static_cast<int>(i); idx++) {
                     length += lines[idx].length();
                     sum += lines[idx].length() * 0.5 * (lines[idx].a_width + lines[idx].b_width);
                     path.polyline.append(lines[idx].a);
@@ -194,7 +194,7 @@ static ExtrusionPaths thick_polyline_to_extrusion_paths_2(const ThickPolyline& t
         double length = lines[start_index].length();
         double sum = lines[start_index].length() * lines[start_index].a_width;
         path.polyline.append(lines[start_index].a);
-        for (int idx = start_index + 1; idx < final_size; idx++) {
+        for (int idx = start_index + 1; idx < static_cast<int>(final_size); idx++) {
             length += lines[idx].length();
             sum += lines[idx].length() * lines[idx].a_width;
             path.polyline.append(lines[idx].a);
